Reworks GOTS
adds WIP seasons attacks

cool stuff!

removes visual clutter

+++ code/modules/mob/living/simple_animal/abnormality/aleph/seasons.dm
@@ -32,6 +32,7 @@
 	is_flying_animal = FALSE
 	start_qliphoth = 1
 	can_breach = TRUE
+	melee_reach = 3
 	work_chances = list(
 		ABNORMALITY_WORK_INSTINCT = list(5, 10, 15, 50, 55),
 		ABNORMALITY_WORK_INSIGHT = list(5, 10, 15, 50, 55),
@@ -115,12 +116,12 @@
 		"winter" = list(RED_DAMAGE = 1.5, WHITE_DAMAGE = 1, BLACK_DAMAGE = 0.8, PALE_DAMAGE = 0.2)
 		)
 
-	//Work Vars
+	// Work Vars
 	var/current_season = "winter"
 	var/downgraded
 	var/safe
 	var/work_timer
-	//Breach Vars
+	// Breach Vars
 	var/can_act = TRUE
 	var/slam_damage = 200
 	var/slam_cooldown
@@ -131,8 +132,11 @@
 	var/pulse_cooldown
 	var/pulse_cooldown_time = 4 SECONDS
 	var/pulse_damage = 15
+	var/special_melee_attack = null
+	// Turf Tracker
+	var/list/spawned_turfs = list()
 
-	//PLAYABLES ATTACKS
+	// PLAYABLES ATTACKS
 	attack_action_types = list(
 		/datum/action/cooldown/seasons_slam,
 		/datum/action/innate/abnormality_attack/toggle/seasons_cone_toggle,
@@ -228,10 +232,10 @@
 			return chance + 35
 	return chance
 
-//Transformations
+// Transformations
 /mob/living/simple_animal/hostile/abnormality/seasons/proc/Transform()
 	current_season = SSlobotomy_events.current_season
-	ChangeResistances(modular_damage_coeff[current_season])
+	ChangeResistances(modular_damage_coeff[current_season]) // So basically what we're doing here is grabbing all of the stats from various associated lists
 	work_damage_type = season_stats[current_season][2]
 	melee_damage_type = season_stats[current_season][2]
 	icon_state = current_season
@@ -242,6 +246,18 @@
 	projectilesound = breaching_stats[current_season][1]
 	playsound(get_turf(src), "[breaching_stats[current_season][2]]", 30, 0, 8)
 	projectiletype = breaching_stats[current_season][3]
+	switch(current_season)
+		if("spring")
+			special_melee_attack = PROC_REF(SpringMelee) // I initially wanted to do these with lists like everything else, but BYOND didn't like it.
+		if("summer")
+			special_melee_attack = PROC_REF(SummerMelee)
+		if("fall")
+			special_melee_attack = PROC_REF(FallMelee)
+		if("winter")
+			special_melee_attack = PROC_REF(WinterMelee)
+			cone_attack_damage = 75
+			slam_damage = 125
+			pulse_damage = 10
 	if(datum_reference != null) // Do not change containment conditions if it does not have a containment cell assigned.
 		var/list/new_work_chances = modular_work_chance[current_season]
 		work_chances = new_work_chances.Copy()
@@ -251,11 +267,7 @@
 		portrait = "[current_season]"
 		name = season_stats[current_season][3]
 		desc = season_stats[current_season][5]
-	if(current_season == "winter")
-		cone_attack_damage = 75
-		slam_damage = 125
-		pulse_damage = 10
-	else
+	if(current_season != "winter")
 		cone_attack_damage = initial(cone_attack_damage)
 		slam_damage = initial(slam_damage)
 		pulse_damage = initial(pulse_damage)
@@ -288,7 +300,7 @@
 	REMOVE_TRAIT(src, TRAIT_MOVE_FLYING, INNATE_TRAIT)
 	update_icon()
 	for(var/turf/open/O in range(1, src))
-		new /obj/effect/season_turf(O)
+		new /obj/effect/season_turf(O) // We create this manually instead of using the proc so it is never deleted.
 
 //Breach Stuff
 /mob/living/simple_animal/hostile/abnormality/seasons/proc/WorkCheck()
@@ -316,11 +328,16 @@
 		StartWeather()
 
 /mob/living/simple_animal/hostile/abnormality/seasons/death(gibbed)
-	EndWeather()
 	density = FALSE
 	animate(src, alpha = 0, time = 10 SECONDS)
 	QDEL_IN(src, 10 SECONDS)
 	..()
+
+/mob/living/simple_animal/hostile/abnormality/seasons/Destroy()
+	EndWeather()
+	for(var/obj/effect/season_turf/newturf in spawned_turfs)
+		newturf.DoDelete()
+	return ..()
 
 /mob/living/simple_animal/hostile/abnormality/seasons/BreachEffect(mob/living/carbon/human/user, breach_type)
 	if(downgraded)
@@ -358,7 +375,11 @@
 	if(ishuman(attacked_target))
 		if(Finisher(attacked_target))
 			return
-	return ..()
+	if(special_melee_attack) // Use a season-specific special melee attack. Notice how we dont call the parent here!
+		if(attacked_target)
+			. = call(src, special_melee_attack)(attacked_target)
+		else
+			. = call(src, special_melee_attack)()
 
 /mob/living/simple_animal/hostile/abnormality/seasons/OpenFire()
 	if(!can_act)
@@ -392,11 +413,7 @@
 	if(!can_act)
 		return FALSE
 	for(var/turf/open/O in range(1, src))
-		if(!isturf(O) || isspaceturf(O))
-			continue
-		if(locate(/obj/effect/season_turf/temporary) in O)
-			continue
-		new /obj/effect/season_turf/temporary(O)
+		TryCreateSeasonTurf(O)
 	..()
 
 
@@ -433,7 +450,9 @@
 	for(var/turf/T in turfs)
 		if(istype(T, /turf/closed))
 			break
-		new attacktype(T)
+		var/obj/effect/season_effect/the_attack = new attacktype(T)
+		if (istype(the_attack, /obj/effect/season_effect))
+			the_attack.source = src
 		for(var/mob/living/L in T.contents)
 			if(L in hit_list || istype(L, type))
 				continue
@@ -459,7 +478,9 @@
 	for(var/turf/T in view(7, src))
 		if((get_dist(src, T) % 2 != 1))
 			continue
-		new attacktype(T)
+		var/obj/effect/season_effect/the_attack = new attacktype(T)
+		if (istype(the_attack, /obj/effect/season_effect))
+			the_attack.source = src
 		for(var/mob/living/L in T)
 			if(faction_check_mob(L))
 				continue
@@ -488,8 +509,7 @@
 	new attacktype(T)
 
 /mob/living/simple_animal/hostile/abnormality/seasons/proc/PulseHit(turf/T)
-	if(!locate(/obj/effect/season_turf/temporary) in T)
-		new /obj/effect/season_turf/temporary(T)
+	TryCreateSeasonTurf(T)
 	for(var/mob/living/L in T)
 		if(faction_check_mob(L))
 			continue
@@ -518,6 +538,106 @@
 				H.Drain()
 		return TRUE
 	return FALSE
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/SpringMelee(atom/attacked_target)
+	to_chat(world, "SpringMelee attack used")
+	return
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/SummerMelee(atom/attacked_target)
+	can_act = FALSE
+	changeNext_move(10) //Prevents attack spam
+	var/turf/TT = get_turf(get_step(src, dir))
+	var/turf/T = get_turf(src)
+	var/current_orientation = pick(1, -1) // Makes it so AOE is flipped randomly
+	var/rotate_dir = current_orientation
+	var/angle_to_target = Get_Angle(T, TT)
+	var/angle = angle_to_target + (300 * rotate_dir) * 0.5
+	if(angle > 360)
+		angle -= 360
+	else if(angle < 0)
+		angle += 360
+	var/turf/T2 = get_turf_in_angle(angle, T, 4)
+	var/list/line = getline(T, T2)
+	for(var/i = 1 to 20)
+		angle += ((300 / 20) * rotate_dir)
+		if(angle > 360)
+			angle -= 360
+		else if(angle < 0)
+			angle += 360
+		T2 = get_turf_in_angle(angle, T, 4)
+		line = getline(T, T2)
+		addtimer(CALLBACK(src, PROC_REF(DoLineWarning), line, TT, src), i * 0.12)
+		addtimer(CALLBACK(src, PROC_REF(DoLineAttack), line, TT, src), i * 0.12 + 8)
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/DoLineWarning(list/line, atom/target, mob/living/carbon/human/user)
+	playsound(src, 'sound/abnormalities/seasons/summer_change.ogg', 30, TRUE, 3)
+	for(var/turf/T in line)
+		if(locate(/obj/effect/temp_visual/cult/sparks) in T)
+			continue
+		new /obj/effect/temp_visual/cult/sparks(T)
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/DoLineAttack(list/line, atom/target, mob/living/carbon/human/user)
+	var/list/been_hit = list()
+	for(var/turf/T in line)
+		if(locate(/obj/effect/temp_visual/smash_effect) in T)
+			continue
+		playsound(T, 'sound/weapons/fixer/generic/fire3.ogg', 30, TRUE, 3)
+		new /obj/effect/temp_visual/smash_effect(T)
+		new /obj/effect/temp_visual/fire/fast(T)
+		been_hit = HurtInTurf(T, been_hit, (melee_damage_upper * 0.25), BURN, check_faction = TRUE)
+		been_hit = HurtInTurf(T, been_hit, melee_damage_upper, RED_DAMAGE, check_faction = TRUE)
+		TryCreateSeasonTurf(T)
+	can_act = TRUE
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/FallMelee(atom/attacked_target)
+	to_chat(world, "FallMelee attack used")
+	return
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/WinterMelee(atom/attacked_target)
+	can_act = FALSE
+	changeNext_move(10) //Prevents attack spam
+	var/list/target_list = list()
+	for(var/mob/living/L in urange(10, src))
+		if(L.z != z || (L.status_flags & GODMODE))
+			continue
+		if(faction_check_mob(L, FALSE))
+			continue
+		if(L.stat == DEAD)
+			continue
+		target_list += L
+	var/projectile_number = (length(target_list) + 3)
+	for(var/i = 1 to projectile_number)
+		if(LAZYLEN(target_list))
+			target = pick(target_list)
+		if(!target)
+			return
+		var/turf/T = get_step(get_turf(src), pick(1,2,4,5,6,8,9,10))
+		if(T.density)
+			i -= 1
+			continue
+		var/obj/projectile/season_projectile/winter/weak/P
+		P = new(T)
+		P.starting = T
+		P.firer = src
+		P.fired_from = T
+		P.yo = target.y - T.y
+		P.xo = target.x - T.x
+		P.original = target
+		P.preparePixelProjectile(target, T)
+		addtimer(CALLBACK (P, TYPE_PROC_REF(/obj/projectile, fire)), i + 6)
+	playsound(get_turf(src), 'sound/abnormalities/seasons/winter_change.ogg', 15, 0, 2)
+	SLEEP_CHECK_DEATH(7)
+	playsound(get_turf(src), 'sound/abnormalities/seasons/winter_attack.ogg', 50, 0, 4)
+	SLEEP_CHECK_DEATH(3)
+	can_act = TRUE
+
+/mob/living/simple_animal/hostile/abnormality/seasons/proc/TryCreateSeasonTurf(turf/open/O)
+	if(!isturf(O) || isspaceturf(O))
+		return
+	if(locate(/obj/effect/season_turf) in O)
+		return
+	var/obj/effect/season_turf/newturf = new(O)
+	spawned_turfs += newturf
 
 //Weather and such
 /datum/weather/thunderstorm //Spring weather, might want to add thunder strikes or make it a bit more dangerous overall.
@@ -582,7 +702,7 @@
 		if(prob(10))
 			if(prob(66))
 				sleep(rand(1,5))
-			new /obj/structure/turf_fire(T)
+			new /obj/effect/turf_fire(T)
 
 /datum/weather/fog //Fall weather, causes nearsightedness.
 	name = "fog"
@@ -696,7 +816,7 @@
 	variable = TRUE
 
 //Misc. Objects
-/obj/effect/season_turf //Modular turf that spawnes under the abnormality with Upgrade(). This version never despawns.
+/obj/effect/season_turf //Modular turf that spawnes under the abnormality with Upgrade().
 	name = "grass"
 	desc = "A thick layer of foilage that never seems to die down."
 	icon = 'icons/turf/floors.dmi'
@@ -710,6 +830,12 @@
 		"winter" = list("snow","A patch of snow."),
 	)
 	var/current_season
+	var/area_affected
+	var/damaging
+	var/list/viable_landmarks = list(
+		/obj/effect/landmark/xeno_spawn,
+		/obj/effect/landmark/department_center
+		)
 
 /obj/effect/season_turf/Initialize()
 	. = ..()
@@ -721,22 +847,36 @@
 	icon = 'icons/turf/floors.dmi'
 	name = season_list[current_season][1]
 	desc = season_list[current_season][2]
+	for(var/X in GLOB.department_centers) // It would make sense to check xeno spawns in an area, but that actually loops through every turf in the world.
+		var/turf/T = X
+		if(get_dist(T, src) < 4)
+			area_affected = TRUE
+			break
+	if(!area_affected)
+		for(var/Y in GLOB.xeno_spawn)
+			var/turf/F = Y
+			if(get_dist(F, src) < 2)
+				area_affected = TRUE
+				break
 	switch(current_season)
 		if("spring")
+			if(area_affected)
+				icon_state = "fairygrass[rand(0,3)]"
+				return
 			icon_state = "grass[rand(0,3)]"
 		if("summer")
-			if(prob(15))
+			if(area_affected)
 				icon_state = "lava"
 				return
 			icon_state = "basalt[rand(0,3)]"
 		if("fall")
-			if(prob(50))
+			if(area_affected)
 				icon_state = "junglegrass"
 				return
 			icon_state = "wasteland[rand(8,11)]"
 		if("winter")
 			icon = 'icons/turf/snow.dmi'
-			if(prob(15))
+			if(area_affected)
 				icon_state = "ice"
 				name = "ice"
 				desc = "A patch of slippery ice."
@@ -751,37 +891,53 @@
 	BumpEffect(H)
 
 /obj/effect/season_turf/proc/BumpEffect(mob/living/carbon/human/H)
+	if(!area_affected)
+		return
 	switch(current_season)
 		if("spring")
-			if(prob(5))
-				to_chat(H, span_warning("Your legs are cut by brambles in the grass!"))
-				H.apply_damage(5, BLACK_DAMAGE, null, H.run_armor_check(null, WHITE_DAMAGE), spread_damage = FALSE)
+			to_chat(H, span_warning("You are stung by nettles as you pass through!"))
+			DoDamage()
 		if("summer")
-			if(icon_state == "lava")
-				to_chat(H, span_warning("You stumbled into a pool of lava!"))
-				H.adjust_fire_stacks(rand(0.1, 1))
-				H.IgniteMob()
+			to_chat(H, span_warning("You stumbled into a pool of lava!"))
+			DoDamage()
 		if("fall")
-			if(prob(5))
-				to_chat(H, span_warning("You sink into the marsh!"))
-				animate(H, alpha = 255,pixel_x = 0, pixel_z = -3, time = 0.5 SECONDS)
-				H.pixel_z = -3
-				H.Immobilize(0.5 SECONDS)
-				animate(H, alpha = 255,pixel_x = 0, pixel_z = 3, time = 0.5 SECONDS)
-				H.pixel_z = 0
+			to_chat(H, span_warning("You sink into the marsh!"))
+			animate(H, alpha = 255,pixel_x = 0, pixel_z = -3, time = 0.5 SECONDS)
+			H.pixel_z = -3
+			H.Immobilize(0.5 SECONDS)
+			animate(H, alpha = 255,pixel_x = 0, pixel_z = 3, time = 0.5 SECONDS)
+			H.pixel_z = 0
 		if("winter")
-			if(icon_state == "ice")
-				if(prob(25))
-					to_chat(H, span_warning("You slip on the ice!"))
-					H.slip(0, null, SLIDE_ICE, 0, FALSE) //might need to replace this as stuns are pretty annoying...
-					H.Immobilize(0.5 SECONDS)
+			if(H.body_position == LYING_DOWN)
+				step(H, H.dir)
+				return
+			if(prob(25))
+				to_chat(H, span_warning("You lose your footing on the ice!"))
+				H.apply_lc_tremor(1, 3)
+				step(H, H.dir)
+				return
+			to_chat(H, span_notice("You manage to keep your balance on the slippery ice."))
 
-/obj/effect/season_turf/temporary //Season_turf but despawns, spawned by most of the abnormality's attacks.
+/obj/effect/season_turf/proc/DoDamage()
+	var/dealt_damage = FALSE
+	for(var/mob/living/L in get_turf(src))
+		if(ishuman(L))
+			var/mob/living/carbon/human/H = L
+			if(current_season == "summer")
+				H.deal_damage(6, FIRE)
+				H.apply_lc_burn(3)
+				dealt_damage = TRUE
+			else if(current_season == "spring")
+				H.apply_damage(10, WHITE_DAMAGE, null, H.run_armor_check(null, WHITE_DAMAGE), spread_damage = FALSE)
+	if(!dealt_damage)
+		damaging = FALSE
+		return
+	addtimer(CALLBACK(src, PROC_REF(DoDamage)), 4)
 
-/obj/effect/season_turf/temporary/Initialize()
-	. = ..()
-	animate(src, alpha = 150, time = 10 SECONDS)
-	QDEL_IN(src, 10 SECONDS)
+/obj/effect/season_turf/proc/DoDelete()
+	var/randomtimer = rand(3,15)
+	animate(src, alpha = 150, time = randomtimer SECONDS)
+	QDEL_IN(src, randomtimer SECONDS)
 
 //Effects
 /obj/effect/season_warn //warning for attacks
@@ -823,23 +979,29 @@
 	name = "weather warning"
 	desc = "Watch out!"
 	layer = POINT_LAYER
+	var/mob/living/simple_animal/hostile/abnormality/seasons/source
 
 /obj/effect/season_effect/Initialize()
 	. = ..()
 	addtimer(CALLBACK(src, PROC_REF(pop)), 0.5 SECONDS)
 
 /obj/effect/season_effect/proc/pop()
-	if(!locate(/obj/effect/season_turf/temporary) in get_turf(src))
-		new /obj/effect/season_turf/temporary(get_turf(src))
+	if(!source)
+		QDEL_IN(src, 0.5 SECONDS)
+		return
+	if(!locate(/obj/effect/season_turf) in get_turf(src))
+		var/obj/effect/season_turf/newturf = new(get_turf(src))
+		source.spawned_turfs += newturf
 	if(prob(5))
 		var/list/spawn_area = range(1, get_turf(src))
 		for(var/turf/open/O in spawn_area)
 			if(!isturf(O) || isspaceturf(O))
 				continue
-			var/obj/effect/season_turf/temporary/G = (locate(/obj/effect/season_turf/temporary) in O)
+			var/obj/effect/season_turf/G = (locate(/obj/effect/season_turf) in O)
 			if(G)
 				qdel(G)
-			new /obj/effect/season_turf/temporary(O)
+			var/obj/effect/season_turf/anewturf = new(O)
+			source.spawned_turfs += anewturf
 	QDEL_IN(src, 0.5 SECONDS)
 
 /obj/effect/season_effect/summer
@@ -864,7 +1026,8 @@
 
 /obj/effect/season_effect/breath/pop()
 	if(prob(75))
-		new /obj/effect/season_turf/temporary(get_turf(src))
+		..()
+		return
 	QDEL_IN(src, 0.5 SECONDS)
 
 /obj/effect/season_effect/breath/summer
